requirements.txt
streamlit
folium
streamlit-folium
plotly
pandas
fpdf2
pillow
